Move JVM utilities service implementation in its own package
