+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/internal/support/package-info.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * This package contains a number of utilities used to facilitate the
+ * integration of Java plugins. It provides the implementation of the
+ * {@link org.gradle.api.plugins.JvmEcosystemUtilities} service.
+ */
+@NonNullApi
+package org.gradle.api.plugins.internal.support;
+
+import org.gradle.api.NonNullApi;
